refactor: рефактор страницы Articles по замечаниям
