+++ src/pages/Articles/Articles.js
@@ -1,7 +1,6 @@
-import './Articles.scss';
 import { useContext, useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import articlesPageTexts from '../../locales/articles-page-RU';
+import articlesPageTexts from './locales/RU';
 import {
   COLORS,
   ERROR_MESSAGES,
@@ -15,14 +14,17 @@ import {
   Paginate,
   TitleH1,
 } from './index';
-import { getArticlesPageData, getArticle } from '../../api/articles-page';
+import { getArticle, getArticlesPageData } from '../../api/articles-page';
 import { setLocalStorageData } from '../../hooks/useLocalStorage';
 import { PopupsContext } from '../../contexts';
+import './Articles.scss';
 
 const PAGE_SIZE_PAGINATE = {
   small: 8,
   big: 12,
 };
+
+const smallQueryWidth = '1024px';
 
 const { headTitle, headDescription, title, textStubNoData } = articlesPageTexts;
 
@@ -42,7 +44,7 @@ function Articles() {
   const searchArticleId = state?.id;
   const { openPopupArticle } = useContext(PopupsContext);
 
-  function getPageData() {
+  const getPageData = () => {
     const offset = pageSize * pageNumber;
     const fixedPageSize =
       pageNumber === 0 && mainArticle ? pageSize + 1 : pageSize;
@@ -60,9 +62,9 @@ function Articles() {
       .finally(() => {
         setIsLoadingPaginate(false);
       });
-  }
+  };
 
-  // Откртие попапа при переходе из поиска
+  // Открытие попапа при переходе из поиска
   useEffect(() => {
     if (state) {
       getArticle(searchArticleId)
@@ -110,7 +112,7 @@ function Articles() {
   }, [pageSize]);
 
   useEffect(() => {
-    const smallQuery = window.matchMedia('(max-width: 1024px)');
+    const smallQuery = window.matchMedia(`(max-width: ${smallQueryWidth})`);
 
     const listener = () => {
       if (smallQuery.matches) {
@@ -128,7 +130,38 @@ function Articles() {
     };
   }, []);
 
-  // отрисовка заглушки
+  if (isLoadingPage) {
+    return <Loader isCentered />;
+  }
+
+  return (
+    <BasePage headTitle={headTitle} headDescription={headDescription}>
+      {!articlesPageData && !isLoadingPage
+        ? renderAnimatedContainer()
+        : renderPageContent()}
+    </BasePage>
+  );
+
+  function renderPageContent() {
+    if (isPageError) {
+      return (
+        <AnimatedPageContainer
+          titleText={ERROR_MESSAGES.generalErrorMessage.title}
+        />
+      );
+    }
+
+    return (
+      <section className="articles page__section">
+        <TitleH1 title={title} sectionClass="fade-in" />
+
+        {isLoadingPaginate ? <Loader isPaginate /> : renderCards()}
+
+        {renderPagination()}
+      </section>
+    );
+  }
+
   function renderAnimatedContainer() {
     if (isPageError) {
       return (
@@ -176,38 +209,6 @@ function Articles() {
       </>
     );
   }
-
-  function renderPageContent() {
-    if (isPageError) {
-      return (
-        <AnimatedPageContainer
-          titleText={ERROR_MESSAGES.generalErrorMessage.title}
-        />
-      );
-    }
-
-    return (
-      <section className="articles page__section">
-        <TitleH1 title={title} sectionClass="fade-in" />
-
-        {isLoadingPaginate ? <Loader isPaginate /> : renderCards()}
-
-        {renderPagination()}
-      </section>
-    );
-  }
-
-  if (isLoadingPage) {
-    return <Loader isCentered />;
-  }
-
-  return (
-    <BasePage headTitle={headTitle} headDescription={headDescription}>
-      {!articlesPageData && !isLoadingPage
-        ? renderAnimatedContainer()
-        : renderPageContent()}
-    </BasePage>
-  );
 }
 
 export default Articles;
