+++ src/pages/Articles/index.js
@@ -1,7 +1,7 @@
 import BasePage from '../../layout/BasePage';
 import CardArticle from '../../components/Cards/CardArticle/CardArticle';
 import AnimatedPageContainer from '../../components/AnimatedPageContainer/AnimatedPageContainer';
-import { TitleH1, Loader, Paginate } from '../../components/utils/index';
+import { Loader, Paginate, TitleH1 } from '../../components/utils';
 
 export {
   BasePage,
