+++ src/pages/Articles/locales/RU.js
@@ -1,4 +1,4 @@
-import { ARTICLES_TITLE, MAIN_PAGE_TITLE } from '../config/routes';
+import { ARTICLES_TITLE, MAIN_PAGE_TITLE } from '../../../config/routes';
 
 const articlesPageTexts = {
   headTitle: `${MAIN_PAGE_TITLE} | ${ARTICLES_TITLE}`,
